520896: Need a way to update existing installation through MPC
Fixed update check for entry IUs

Update information was incomplete for entries from drag&drop and for
entries that were not considered "installed" as a whole because
required items were missing, but that had some IUs already installed.
This could lead to wrong operation states on the feature selection
page if one of them had an update available.

Bug: 520896
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=520896