Bug 536589: Add support of new uriScheme extension point
This adds an uri scheme handler for the custom uri scheme
"eclipse-mpc".

Marketplace client currently get's URLs in the form
http://marketplace.eclipse.org/marketplace-client-intro?mpc_install=311881
"eclipse-mpc"-URLs are exactly the same only the "http" is exchanged
with "eclipse.-mpc". The idea is that the marketplace website renders
links with the "eclipse-mpc" URI scheme.

The implementation is quite easy. It just replaces the "eclipse-mpc"
back to "https" and just calls the existing facilities in the
MarketplaceUrlHandler class.

Change-Id: I6934c35e0188288ffa6962bf0d6f62e369e9a05c
Signed-off-by: Matthias Becker <ma.becker@sap.com>