Bug 568414 - Remove Activator from MPC Core
The MPC.core plugin uses a Bundle-Activator which sets up some internal
state, but most of this can be refactored away. This allows applications
that depend on MPC to start faster.

Refactor dependencies that are present in the
MarketplaceClientCorePlugin so that the start/stop methods are no longer
required.

In addition, bump the require execution environment to Java 11, along
with a minor service bump, since it depends on bundles that depend on
Java 11.

Change-Id: Idbb5d0205e3f2864db31fef61345779508cab68d
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>