Bug 568466 - Remove reference to Bundle
The bundle is only stored to access the symbolic name for the purposes
of image lookup, and that's not necessary because there's an alternate
function which passes in a class from that bundle which is more
efficient. Replace calls to the bundle's symbolic name with the class
and remove caching the bundle.

Change-Id: I6f98a680f7cac65ff5d4d6f48470011fd6e26879
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>