Bug 568721 Remove MarketPlaceClientUi activator
The Bundle-Activator is only used to stop the jobs associated with this
bundle; but the only one that looks out for the `belongsTo` call is the
`MarketPlaceDiscoveryStrategy` which doesn't check for cancellation via
the passed `IProgressMonitor`. As a result, it will run to completion
regardless of whether it's cancelled or not.

Once this call becomes unnecessary, we can remove the Bundle-Activator
from the manifest completely.

Change-Id: I586669078516aa1a382a97a5cd24da0732936a1f
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>